pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-contract-validator"
version = "0.1.0"
description = "Multi-dimensional API contract validator with intelligent test generation and drift detection"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["api", "contract", "validation", "testing", "openapi", "swagger", "drift-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "openapi-spec-validator>=0.7.0,<1.0.0",
    "pyyaml>=6.0,<7.0",
    "jsonschema>=4.0.0,<5.0.0",
    "httpx>=0.27.0,<1.0.0",
    "click>=8.1.0,<9.0.0",
    "rich>=13.0.0,<14.0.0",
    "jinja2>=3.1.0,<4.0.0",
    "spacy>=3.7.0,<4.0.0",
    "anthropic>=0.25.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "tinydb>=4.8.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-httpserver>=1.0.0,<2.0.0",
    "black>=24.0.0,<25.0.0",
    "ruff>=0.5.0,<1.0.0",
    "mypy>=1.10.0,<2.0.0",
    "types-pyyaml>=6.0.0",
    "types-jsonschema>=4.0.0",
]

[project.scripts]
api-contract-validator = "api_contract_validator.cli.main:cli"
acv = "api_contract_validator.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourusername/api-contract-validator"
Documentation = "https://github.com/yourusername/api-contract-validator/docs"
Repository = "https://github.com/yourusername/api-contract-validator"
Issues = "https://github.com/yourusername/api-contract-validator/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["api_contract_validator*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # Allow raising exceptions without from e
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "openapi_spec_validator.*",
    "tinydb.*",
    "spacy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=api_contract_validator",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
